create motor struct
This wrapper around the motor pins and the dac allows for much simpler
and less verbose interaction with the peripherals related to the motors.
Instead of having to lock each peripheral's mutex and pass each pin into
functions that require the motor interface, we can now simply lock and
use the Motor struct. Additionally, to change the mode of the motor,
simply pass in a MotorMode enum variant for whichever mode you'd like to
set it to.