Make flashing drogue vs main board easier
Instead of hard coding the environment variable in .cargo/config.toml,
you can now specify the BOARD environment variable through a commmand
line override. This was previously not possible because setting it
through the command line wouldnt trigger a rerun of the build. Adjusting
the build.rs file fixed this issue. I also wrote cargo aliases to
simplify the flash commands, and updated the READMEs.