app: settings: ERS setting module now paired down to single setting . . .
Also find that setting Kconfig symbols serve as needed when present in primary
prj.conf Kconfig file.  Same symbols however are somehow not picked up in time
or effectively when they're factored into a Kconfig overlay in the boards
directory.  Strange, as these symbos are so factored in Zephyr's sample app.